refactor(analysis-types): split churn DTOs
